(chore) test: split MatcherTests into focused test classes
Split the monolithic MatcherTests (4369 lines) into 10 focused test
classes plus a shared helper utility:

- MatcherUnicodeTests: Unicode encoding (1-4 byte, region)
- MatcherMatchingTests: matches(), lookingAt(), find(), capture groups
- MatcherReplacementTests: replaceAll(), replaceFirst(), appendReplacement()
- MatcherResultsTests: results() stream operations
- MatcherAnchoringBoundsTests: anchoring bounds behavior
- MatcherTransparentBoundsTests: transparent/opaque bounds behavior
- MatcherHitEndRequireEndTests: hitEnd(), requireEnd(), empty regions
- MatcherToStringTests: toString() output
- MatcherMatchResultTests: MatchResult snapshot behavior
- MatcherCanonEqTests: CANON_EQ canonical equivalence
- MatcherTestUtils: shared assertion helpers

No test logic changes; purely structural reorganization.
